+++ python/src/scipp/table.py
@@ -246,7 +246,7 @@ class TableViewer:
                 else:
                     self.headers = 0
                     key = " "
-                    var = sc.Variable(['row'], values=scipp_obj)
+                    var = sc.Variable(dims=['row'], values=scipp_obj)
 
                 self.tabledict[group][key] = {"data": {key: var}}
                 self.is_bin_centers[group][key] = {"data": {key: False}}