+++ python/tests/common_test.py
@@ -7,7 +7,7 @@ import scipp as sc
 
 def test_bool_raises():
     # Truth values of arrays are undefined
-    var = sc.Variable(value=True)
+    var = sc.scalar(True)
     with pytest.raises(RuntimeError):
         var and var
     da = sc.DataArray(var)