+++ python/tests/datasetslice_test.py
@@ -9,42 +9,48 @@ from .common import assert_export
 
 class TestDatasetSlice:
     def setup_method(self):
-        var = sc.Variable(['x'], values=np.arange(5, dtype=np.int64))
+        var = sc.Variable(dims=['x'], values=np.arange(5, dtype=np.int64))
         self._d = sc.Dataset(data={'a': var, 'b': var}, coords={'x': var})
 
     def test_slice_with_range_datasetview_then_dataarrayview(self):
         sl = self._d['x', 1:-1]['a'].data
-        ref = sc.Variable(['x'], values=np.array([1, 2, 3], dtype=np.int64))
+        ref = sc.Variable(dims=['x'],
+                          values=np.array([1, 2, 3], dtype=np.int64))
         assert sc.identical(ref, sl)
         # omitting range end
         sl = self._d['x', 1:]['a'].data
-        ref = sc.Variable(['x'], values=np.array([1, 2, 3, 4], dtype=np.int64))
+        ref = sc.Variable(dims=['x'],
+                          values=np.array([1, 2, 3, 4], dtype=np.int64))
         assert sc.identical(ref, sl)
         # omitting range begin
         sl = self._d['x', :-1]['a'].data
-        ref = sc.Variable(['x'], values=np.array([0, 1, 2, 3], dtype=np.int64))
+        ref = sc.Variable(dims=['x'],
+                          values=np.array([0, 1, 2, 3], dtype=np.int64))
         assert sc.identical(ref, sl)
         # omitting range both begin and end
         sl = self._d['x', :]['b'].data
-        ref = sc.Variable(['x'],
+        ref = sc.Variable(dims=['x'],
                           values=np.array([0, 1, 2, 3, 4], dtype=np.int64))
         assert sc.identical(ref, sl)
 
     def test_slice_with_range_dataarrayview_then_dataarrayview(self):
         sl = self._d['a']['x', 1:-1].data
-        ref = sc.Variable(['x'], values=np.array([1, 2, 3], dtype=np.int64))
+        ref = sc.Variable(dims=['x'],
+                          values=np.array([1, 2, 3], dtype=np.int64))
         assert sc.identical(ref, sl)
         # omitting range end
         sl = self._d['a']['x', 1:].data
-        ref = sc.Variable(['x'], values=np.array([1, 2, 3, 4], dtype=np.int64))
+        ref = sc.Variable(dims=['x'],
+                          values=np.array([1, 2, 3, 4], dtype=np.int64))
         assert sc.identical(ref, sl)
         # omitting range begin
         sl = self._d['a']['x', :-1].data
-        ref = sc.Variable(['x'], values=np.array([0, 1, 2, 3], dtype=np.int64))
+        ref = sc.Variable(dims=['x'],
+                          values=np.array([0, 1, 2, 3], dtype=np.int64))
         assert sc.identical(ref, sl)
         # omitting range both begin and end
         sl = self._d['b']['x', :].data
-        ref = sc.Variable(['x'],
+        ref = sc.Variable(dims=['x'],
                           values=np.array([0, 1, 2, 3, 4], dtype=np.int64))
         assert sc.identical(ref, sl)
 
@@ -65,10 +71,12 @@ class TestDatasetSlice:
         N = 6
         M = 4
         d1 = sc.Dataset()
-        d1['x'] = sc.Variable(['x'], values=np.arange(N).astype(np.float64))
-        d1['y'] = sc.Variable(['y'], values=np.arange(M).astype(np.float64))
+        d1['x'] = sc.Variable(dims=['x'],
+                              values=np.arange(N).astype(np.float64))
+        d1['y'] = sc.Variable(dims=['y'],
+                              values=np.arange(M).astype(np.float64))
         arr1 = np.arange(N * M).reshape(N, M).astype(np.float64) + 1
-        d1['a'] = sc.Variable(['x', 'y'], values=arr1)
+        d1['a'] = sc.Variable(dims=['x', 'y'], values=arr1)
         d1 = d1['x', 1:2]
         d1['a'].data.values.tolist() == [[5.0, 6.0, 7.0, 8.0]]
 
@@ -81,7 +89,7 @@ class TestDatasetSlice:
         assert d['a'].data.values.tolist() == [0, 2, 2, 3, 5]
 
     def test_slice_and_dimensions_items_dataarray(self):
-        var = sc.Variable(['x', 'y'], values=np.arange(50).reshape(5, 10))
+        var = sc.Variable(dims=['x', 'y'], values=np.arange(50).reshape(5, 10))
         da = sc.DataArray(var)
         assert np.allclose(da['x', 0].values, da['x', 0:1].values)
         assert np.allclose(da['x', 4].values, da['x', -1].values)
@@ -93,7 +101,7 @@ class TestDatasetSlice:
 
     def test_slice_and_dimensions_items_dataset(self):
         da = sc.DataArray(
-            sc.Variable(['x', 'y'], values=np.arange(50).reshape(5, 10)))
+            sc.Variable(dims=['x', 'y'], values=np.arange(50).reshape(5, 10)))
         ds = sc.Dataset(data={'a': da})
         assert (np.allclose(ds['x', 0]['a'].values,
                             ds['x', 0:1]['a'].values[0],
@@ -107,24 +115,24 @@ class TestDatasetSlice:
 
     def test_slice_dataset_with_data_only(self):
         d = sc.Dataset()
-        d['data'] = sc.Variable(['y'], values=np.arange(10))
+        d['data'] = sc.Variable(dims=['y'], values=np.arange(10))
         sliced = d['y', :]
         assert sc.identical(d, sliced)
         sliced = d['y', 2:6]
-        assert sc.identical(sc.Variable(['y'], values=np.arange(2, 6)),
+        assert sc.identical(sc.Variable(dims=['y'], values=np.arange(2, 6)),
                             sliced['data'].data)
 
     def test_slice_dataset_with_coords_only(self):
         d = sc.Dataset(
-            coords={'y-coord': sc.Variable(['y'], values=np.arange(10))})
+            coords={'y-coord': sc.Variable(dims=['y'], values=np.arange(10))})
         sliced = d['y', :]
         assert sc.identical(d, sliced)
         sliced = d['y', 2:6]
-        assert sc.identical(sc.Variable(['y'], values=np.arange(2, 6)),
+        assert sc.identical(sc.Variable(dims=['y'], values=np.arange(2, 6)),
                             sliced.coords['y-coord'])
 
     def test_slice_with_step_1(self):
-        var = sc.Variable(['x'], values=np.arange(1, 4, dtype=np.int64))
+        var = sc.Variable(dims=['x'], values=np.arange(1, 4, dtype=np.int64))
         expect = sc.Dataset(data={'a': var, 'b': var}, coords={'x': var})
         assert sc.identical(self._d['x', 1:4:1], expect)
         assert sc.identical(self._d['a']['x', 1:4:1], expect['a'])