+++ python/tests/variable_comparison_test.py
@@ -17,32 +17,32 @@ def _check_comparison_ops_on(obj):
 
 
 def test_comparison_op_exports_for_variable():
-    var = sc.Variable(['x'], values=np.array([1, 2, 3]))
+    var = sc.Variable(dims=['x'], values=np.array([1, 2, 3]))
     _check_comparison_ops_on(var)
     _check_comparison_ops_on(var['x', :])
 
 
 def test_isclose():
     unit = sc.units.one
-    a = sc.Variable(['x'], values=np.array([1, 2, 3]), unit=unit)
+    a = sc.Variable(dims=['x'], values=np.array([1, 2, 3]), unit=unit)
     assert sc.all(sc.isclose(a, a, rtol=0 * unit, atol=0 * unit)).value
 
 
 def test_isclose_atol_defaults():
     unit = sc.units.one
-    a = sc.Variable(['x'], values=np.array([1, 2, 3]), unit=unit)
+    a = sc.Variable(dims=['x'], values=np.array([1, 2, 3]), unit=unit)
     assert sc.all(sc.isclose(a, a, rtol=0 * unit)).value
 
 
 def test_isclose_rtol_defaults():
     unit = sc.units.one
-    a = sc.Variable(['x'], values=np.array([1, 2, 3]), unit=unit)
+    a = sc.Variable(dims=['x'], values=np.array([1, 2, 3]), unit=unit)
     assert sc.all(sc.isclose(a, a, atol=0 * unit)).value
 
 
 def test_allclose():
     unit = sc.units.one
-    a = sc.Variable(['x'], values=np.array([1, 2, 3]), unit=unit)
+    a = sc.Variable(dims=['x'], values=np.array([1, 2, 3]), unit=unit)
     assert sc.allclose(a, a, 0 * unit, 0 * unit)
     b = a.copy()
     b['x', 0] = 2
@@ -57,18 +57,18 @@ def test_allclose_vectors():
 
 def test_allclose_atol_defaults():
     unit = sc.units.one
-    a = sc.Variable(['x'], values=np.array([1, 2, 3]), unit=unit)
+    a = sc.Variable(dims=['x'], values=np.array([1, 2, 3]), unit=unit)
     assert sc.allclose(a, a, rtol=0 * unit)
 
 
 def test_allclose_rtol_defaults():
     unit = sc.units.one
-    a = sc.Variable(['x'], values=np.array([1, 2, 3]), unit=unit)
+    a = sc.Variable(dims=['x'], values=np.array([1, 2, 3]), unit=unit)
     assert sc.allclose(a, a, atol=0 * unit)
 
 
 def test_identical():
-    var = sc.Variable(['x'], values=np.array([1]))
+    var = sc.Variable(dims=['x'], values=np.array([1]))
     assert_export(sc.identical, var, var)
     assert_export(sc.identical, var['x', :], var['x', :])