+++ python/tests/variable_py_object_test.py
@@ -6,7 +6,7 @@ import scipp as sc
 
 
 def test_scalar_Variable_py_object_dict():
-    var = sc.Variable(value={'a': 1, 'b': 2})
+    var = sc.scalar({'a': 1, 'b': 2})
     assert var.dtype == sc.dtype.PyObject
     assert var.value == {'a': 1, 'b': 2}
     var.value['a'] = 3
@@ -15,7 +15,7 @@ def test_scalar_Variable_py_object_dict():
 
 
 def test_scalar_Variable_py_object_list():
-    var = sc.Variable(value=[1, 2, 3])
+    var = sc.scalar([1, 2, 3])
     assert var.dtype == sc.dtype.PyObject
     assert var.value == [1, 2, 3]
     var.value[0] = 2
@@ -23,7 +23,7 @@ def test_scalar_Variable_py_object_list():
 
 
 def test_scalar_Variable_py_object_change():
-    var = sc.Variable(value=[1, 2, 3])
+    var = sc.scalar([1, 2, 3])
     assert var.dtype == sc.dtype.PyObject
     assert var.value == [1, 2, 3]
     # Value assignment cannot change dtype, so the result is still PyObject,
@@ -35,13 +35,13 @@ def test_scalar_Variable_py_object_change():
 
 def test_scalar_Variable_py_object_shares():
     d = {'a': 1, 'b': 2}
-    var = sc.Variable(value=d)
+    var = sc.scalar(d)
     d['a'] += 1
     assert var.value == {'a': 2, 'b': 2}
 
 
 def test_scalar_Variable_py_object_copy_is_deep_copy():
-    var = sc.Variable(value=[1, 2, 3])
+    var = sc.scalar([1, 2, 3])
     # Dataset.copy() releases the GIL. This will segfault unless
     # scipp::python::PyObject::PyObject acquires the GIL.
     copy = var.copy()
@@ -51,14 +51,14 @@ def test_scalar_Variable_py_object_copy_is_deep_copy():
 
 
 def test_scalar_Variable_py_object_comparison():
-    a = sc.Variable(value=[1, 2])
-    b = sc.Variable(value=[1, 2])
+    a = sc.scalar([1, 2])
+    b = sc.scalar([1, 2])
     assert sc.identical(a, b)
 
 
 def test_py_object_delitem():
     d = sc.Dataset()
-    d['a'] = sc.Variable(value=[1, 2])
+    d['a'] = sc.scalar([1, 2])
     # Dataset.__delitem__ releases the GIL. This will segfault unless
     # scipp::python::PyObject::~PyObject acquires the GIL.
     del d['a']